Added export and run methods to Technology class
Added "export" method to Technology class.
The export method will understand if the current library is required or
called by a command line. In the first case it will return the class
prototype to the nodejs module. Otherwise it will call the "run" method
and execute the Technology.

Added "run" method to Technology class.
When called it will start the technology.